Split the Newsletters.scala file in two:  - The newsletter API that loads the definition from the API  - The newsletter model and its constants
